Move TestHelpers to integration test package
