+++ src/test/java/org/cyclops/integrateddynamics/core/evaluate/variable/integration/TestHelpers.java
@@ -1,4 +1,4 @@
-package org.cyclops.integrateddynamics.core.test;
+package org.cyclops.integrateddynamics.core.evaluate.variable.integration;
 
 import org.apache.http.util.Asserts;
