+++ src/test/java/org/cyclops/integrateddynamics/core/evaluate/variable/integration/TestItemStackOperators.java
@@ -25,7 +25,6 @@ import org.cyclops.integrateddynamics.block.BlockEnergyBatteryConfig;
 import org.cyclops.integrateddynamics.core.evaluate.operator.Operators;
 import org.cyclops.integrateddynamics.core.evaluate.variable.*;
 import org.cyclops.integrateddynamics.core.helper.Helpers;
-import org.cyclops.integrateddynamics.core.test.TestHelpers;
 import org.cyclops.integrateddynamics.item.ItemWrench;
 import org.junit.Before;
 import org.junit.Test;
